requirements.txt
streamlit
pandas
google-auth
google-auth-oauthlib
google-auth-httplib2
gspread
Pillow
